backend/requirements.txt
python-novaclient==9.1.1
python-neutronclient==7.2.1
aiohttp==3.5.4
git+https://github.com/wtsi-hgi/osdataproc.git
